+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # 1.0.0 (2023-06-18)
 
-
 ### Features
 
-* recipe crud operations ([c5921af](https://github.com/julian-one/jubilance/commit/c5921afcc046765322d2561a120ff2521a31973c))
+-   recipe crud operations ([c5921af](https://github.com/julian-one/jubilance/commit/c5921afcc046765322d2561a120ff2521a31973c))